--- a/modules/dnn/src/cuda/vector_traits.hpp
+++ b/modules/dnn/src/cuda/vector_traits.hpp
@@ -8,6 +8,7 @@
 #include <cuda_runtime.h>
 
 #include "types.hpp"
+#include "memory.hpp"
 
 #include "../cuda4dnn/csl/pointer.hpp"
 
@@ -84,6 +85,16 @@ namespace cv { namespace dnn { namespace cuda4dnn { namespace csl { namespace de
     template <class V>
     __device__ void v_load(V& dest, const V* src) {
         dest.raw = src->raw;
+    }
+
+    template <class V>
+    __device__ void v_load_ldg(V& dest, const V& src) {
+        dest.raw = load_ldg(src.raw);
+    }
+
+    template <class V>
+    __device__ void v_load_ldg(V& dest, const V* src) {
+        dest.raw = load_ldg(src->raw);
     }
 
     template <class V>